Do not use ClusterRule as @ClassRule
Logs are always deleted when using ClusterRule as a @ClassRule due to
limitation of the TestDirectory rule.  The problem is that if the
latter is used in a static context cannot see single test failures,
hence it will effectively delete the files on disk. This is
problematic when there are failures in CI and we need logs to
understand what happened for debugging purposes.

This commit fix such problems.