Fix TxPushStrategyConfigIT.twoRoundRobin flakyness
The test would fail whenever there are failure in the replication
attempt.  Indeed since the push factor is on a best effort mode in
reality we could be more far behind on slaves of what the test is
claiming.  This change will try to consider the failed replications
when asserting.  Such change should make the test more reliable.