core-edge: fix deadlock between shared discovery service and listener
Previously the RaftDiscoveryServiceConnector would access the
discovery service when invoked on its topology change callback,
but within a synchronized block while at the same time accessing
the services of the discovery service which might be busy trying
to invoke the callback for some other trigger, leading to a
deadlock.

This is now fixed by the discovery client never explcitly asking
about the topology, but rather only being notified when it changes
and keeping that value.

Also fixes a non-synchronized iteration over the listeners in
the SharedDiscoveryCoreClient.