Fix NPE in ServerBootstrapper
When there are validation failure in the configuration we have loaded
from disk a NPE is thrown when we try to log the exception because the
log is initialized later on.  This commit fixes this problems by
making sure that we have a simple log that simply prints to stdout if
the real log is not build yet.