core-edge: Automatic bootstrapping
Previously there was some special tooling required to bootstrap a cluster
from a seed, but now we no longer require any of it. The long term intention
is to just be able to put a store seed, e.g. from a backup or stand-alone
import, into the initial member of a cluster (or all) and have it working.