Fixes an issue with concurrently reading log during rotation
Problem was observed when reading transaction stream from
LogicalTransactionLog when all of a sudden an IOException saying
that log header couldn't be read was thrown. This happened since
there was a concurrent rotation, log version was incremented before
new log file was fully created. There was a guard for that in
the default log version bridge but the guard only guarded for
the file not being present, not for the header not yet having
been written.

Fix is to detect this state and treat it the same way as a missing
log file for that version. We do this by introducing a specific
IncompleteLogHeaderException (extending IOException) and letting
the ReaderLogVersionBridge catch it.

As a side-effect a lot of states and how recovery and concurrent
readers may parry for different states is now well documented in
PhysicalLogFile.

Another thing was that log readers always opened the files in
read/write mode even when purely reading, which was unnecessary.
This was changed to be specified from the outside.

PhysicalLogFileRotateAndReadRaceIT captures and showcases the issue.