Fix secure caching of authentication info
When caching the result of LDAP authentication, the credentials needs to be
stored in order to perform credentials matching against the cached
authentication info. This is done by hashing the credentials with a random salt.

- Use HashedCredentialsMatcher in LdapRealm
- Adds new setting dbms.security.realms.ldap.authentication_cache_enabled
- Move rolePermissionResolver to PredefinedRolesBuilder

This fixes a serious bug where you could previously authenticate against an
existing user's cached authentication result without supplying correct
credentials.